+++ src/srcp_devices_ddl_gl.rs
@@ -96,7 +96,7 @@ pub struct DdlGL<'a> {
   ///Für welche GL's soll ein Oszi Trigger ausgegeben werden?
   trigger: Vec<u32>,
   ///Und Port für Oszi trigger
-  trigger_port: Option<u16>,
+  trigger_port: Option<u32>,
 }
 
 impl DdlGL<'_> {
